Sphere

+++ src/math/vector_properties.c
@@ -14,7 +14,7 @@
 
 float	vec_len(const t_vec3 vector)
 {
-	return (sqrt(vector.x * vector.x
+	return (sqrtf(vector.x * vector.x
 			+ vector.y * vector.y + vector.z * vector.z));
 }
